Program/requirements.txt
deepmultilingualpunctuation==1.0.1
librosa==0.9.1
numpy==1.21.5
pyannote.pipeline==2.2
pydub==0.25.1
scipy==1.8.0
stanza==1.3.0
vosk==0.3.32
